+++ lib/openmaize/login.ex
@@ -30,8 +30,7 @@ defmodule Openmaize.Login do
     case {login_user(user, password), redirects} do
       {false, false} -> send_error(conn, 401, "Invalid credentials")
       {false, true} -> handle_error(conn, "Invalid credentials")
-      {user, false} -> add_token(conn, user, token_opts, nil)
-      {user, true} -> add_token(conn, user, token_opts, storage)
+      {user, _} -> add_token(conn, user, token_opts, storage)
     end
   end