+++ test/authenticate_test.exs
@@ -33,69 +33,49 @@ defmodule Openmaize.AuthenticateTest do
     |> Authenticate.call([storage: nil])
   end
 
-  test "redirect for expired token", %{exp_token: exp_token} do
+  test "expired token", %{exp_token: exp_token} do
     conn = call("/admin", exp_token, :cookie)
-    assert List.keyfind(conn.resp_headers, "location", 0) ==
-      {"location", "/admin/login"}
-    assert conn.status == 302
+    assert conn.assigns ==  %{current_user: nil}
   end
 
-  test "redirect for token that cannot be used yet", %{nbf_token: nbf_token} do
+  test "token that cannot be used yet", %{nbf_token: nbf_token} do
     conn = call("/admin", nbf_token, :cookie)
-    assert List.keyfind(conn.resp_headers, "location", 0) ==
-      {"location", "/admin/login"}
-    assert conn.status == 302
+    assert conn.assigns ==  %{current_user: nil}
   end
 
   test "correct token stored in cookie", %{user_token: user_token} do
-    conn = call("/", user_token, :cookie) |> send_resp(200, "")
-    assert conn.status == 200
+    conn = call("/", user_token, :cookie)
     assert conn.assigns == %{current_user: %{id: 1, name: "Raymond Luxury Yacht", role: "user"}}
   end
 
-  test "redirect for invalid token stored in cookie", %{user_token: user_token} do
+  test "invalid token stored in cookie", %{user_token: user_token} do
     conn = call("/users", user_token <> "a", :cookie)
-    assert List.keyfind(conn.resp_headers, "location", 0) ==
-      {"location", "/admin/login"}
-    assert conn.status == 302
+    assert conn.assigns ==  %{current_user: nil}
   end
 
   test "correct token stored in sessionStorage", %{user_token: user_token} do
-    conn = call("/", user_token, nil) |> send_resp(200, "")
-    assert conn.status == 200
+    conn = call("/", user_token, nil)
     assert conn.assigns ==  %{current_user: %{id: 1, name: "Raymond Luxury Yacht", role: "user"}}
   end
 
-  test "redirect for invalid token stored in sessionStorage", %{user_token: user_token} do
+  test "invalid token stored in sessionStorage", %{user_token: user_token} do
     conn = call("/users", user_token <> "a", nil)
-    assert List.keyfind(conn.resp_headers, "location", 0) ==
-      {"location", "/admin/login"}
-    assert conn.status == 302
-  end
-
-  test "invalid token for unprotected page", %{user_token: user_token} do
-    conn = call("/", user_token <> "a", :cookie) |> send_resp(200, "")
-    assert conn.status == 200
     assert conn.assigns ==  %{current_user: nil}
   end
 
   test "missing token" do
-    conn = conn(:get, "/") |> Authenticate.call([]) |> send_resp(200, "")
-    assert conn.status == 200
+    conn = conn(:get, "/") |> Authenticate.call([])
     assert conn.assigns == %{current_user: nil}
   end
 
   test "correct token using sha256", %{user_256_token: user_256_token} do
-    conn = call("/", user_256_token, :cookie) |> send_resp(200, "")
-    assert conn.status == 200
+    conn = call("/", user_256_token, :cookie)
     assert conn.assigns == %{current_user: %{id: 1, name: "Raymond Luxury Yacht", role: "user"}}
   end
 
-  test "redirect for invalid token using sha256", %{user_256_token: user_256_token} do
+  test "invalid token using sha256", %{user_256_token: user_256_token} do
     conn = call("/users", user_256_token <> "a", :cookie)
-    assert List.keyfind(conn.resp_headers, "location", 0) ==
-      {"location", "/admin/login"}
-    assert conn.status == 302
+    assert conn.assigns ==  %{current_user: nil}
   end
 
 end